Just remove the entire api_backend_service module.
Its name is incorrect...
